+++ NEWS.md
@@ -3,6 +3,11 @@ title: "Release notes for `matsbyname`"
 output: html_document
 ---
 
+Cite all releases with doi [10.5281/zenodo.5118872](https://doi.org/10.5281/zenodo.5118872), 
+which always resolves to the latest release.
+
+
+
 * `matsbyname::Matrix` now vectorized for lists of `matrix` objects
   in the `data` argument.
 * Policy change: